+++ packages/cli/src/controller/migrate/abis/migrate-abis.controller.ts
@@ -1,0 +1,40 @@
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+import fs from 'fs';
+import path from 'path';
+import {uniqWith} from 'lodash';
+import {DEFAULT_ABI_DIR} from '../../generate-controller';
+import {SubgraphDataSource, SubgraphProject} from '../types';
+
+function extractAllAbiFiles(dataSources: SubgraphDataSource[]): string[] {
+  return uniqWith(
+    dataSources.flatMap((dataSource) =>
+      dataSource.mapping.abis.map((abi) => abi.file).filter((file) => file !== null && file !== undefined)
+    )
+  );
+}
+
+async function copyAbiFilesToTargetPathAsync(abiFiles: string[], targetPath: string): Promise<void> {
+  try {
+    for (const file of abiFiles) {
+      const fileName = path.basename(file);
+      const targetFilePath = path.join(targetPath, fileName);
+      await fs.promises.copyFile(file, targetFilePath);
+    }
+    console.log('ABI files copied successfully.');
+  } catch (error) {
+    console.error('Error copying ABI files:', error);
+  }
+}
+
+export async function migrateAbis(
+  subgraphManifest: SubgraphProject,
+  subgraphDir: string,
+  subqlDir: string
+): Promise<void> {
+  const abiPaths = extractAllAbiFiles(subgraphManifest.dataSources);
+  const resolvedPaths = abiPaths.map((p) => path.join(subgraphDir, p));
+  const targetAbiDir = path.join(subqlDir, DEFAULT_ABI_DIR);
+  await copyAbiFilesToTargetPathAsync(resolvedPaths, targetAbiDir);
+}