+++ packages/cli/src/controller/migrate/index.ts
@@ -3,4 +3,6 @@
 
 export * from './migrate-controller';
 export * from './manifest/migrate-manifest.controller';
+export * from './abis/migrate-abis.controller';
 export * from './schema/migrate-schema.controller';
+export * from './constants';