+++ packages/cli/src/controller/migrate/manifest/migrate-manifest.controller.ts
@@ -7,12 +7,34 @@ import {NETWORK_FAMILY} from '@subql/common';
 import {CommonSubqueryProject} from '@subql/types-core';
 import {upperFirst} from 'lodash';
 import YAML from 'yaml';
-import {SubgraphDataSource, SubgraphProject, SubgraphTemplate} from '../../../types';
-import {renderTemplate} from '../../../utils';
-import {DatasourceKind, TemplateKind} from '../../codegen-controller';
-import {convertEthereumDs, convertEthereumTemplate} from './ethereum';
+import {findRunnerByNetworkFamily, prepareDirPath, renderTemplate} from '../../../utils';
+import {TemplateKind} from '../../codegen-controller';
+import {graphToSubqlNetworkFamily, networkDsConverters, networkTemplateConverters} from '../constants';
+import {getChainIdByNetworkName} from '../migrate-controller';
+import {MigrateDatasourceKind, SubgraphDataSource, SubgraphProject, SubgraphTemplate, ChainInfo} from '../types';
 
-const PROJECT_TEMPLATE_PATH = path.resolve(__dirname, '../template/project.ts.ejs');
+const PROJECT_TEMPLATE_PATH = path.resolve(__dirname, '../../../template/project.ts.ejs');
+
+export function readSubgraphManifest(inputPath: string): SubgraphProject {
+  if (!fs.existsSync(inputPath)) {
+    throw new Error(`Subgraph manifest ${inputPath} is not exist`);
+  }
+  const subgraphManifest: SubgraphProject = YAML.parse(fs.readFileSync(inputPath, 'utf8'));
+  return subgraphManifest;
+}
+
+export function extractNetworkFromManifest(subgraphProject: SubgraphProject): ChainInfo {
+  const firstDsKind = subgraphProject.dataSources[0].kind;
+  const firstDsNetwork = subgraphProject.dataSources[0].network;
+  if (!firstDsKind || !firstDsNetwork) {
+    throw new Error(`Subgraph dataSource kind or network not been found`);
+  }
+  const networkFamily = graphToSubqlNetworkFamily[firstDsKind];
+  if (!networkFamily) {
+    throw new Error(`Corresponding subquery network is not find with subgraph data source kind ${firstDsKind}`);
+  }
+  return {networkFamily, chainId: getChainIdByNetworkName(networkFamily, subgraphProject.dataSources[0].network)};
+}
 
 /**
  * Render subquery project to .ts file
@@ -46,10 +68,9 @@ export async function renderManifest(projectPath: string, project: CommonSubquer
  * @param inputPath file path to subgraph.yaml
  * @param outputPath file path to project.ts
  */
-export async function migrateManifest(network: NETWORK_FAMILY, inputPath: string, outputPath: string) {
-  // currently point to subgraph eth project only, can be expand by network
-  const subgraphManifest: SubgraphProject = YAML.parse(fs.readFileSync(inputPath, 'utf8'));
-  await renderManifest(outputPath, graphManifestToSubqlManifest(network, subgraphManifest));
+export async function migrateManifest(chainInfo: ChainInfo, subgraphManifest: SubgraphProject, outputPath: string) {
+  await prepareDirPath(outputPath, false);
+  await renderManifest(outputPath, graphManifestToSubqlManifest(chainInfo, subgraphManifest));
 }
 
 /**
@@ -57,20 +78,22 @@ export async function migrateManifest(network: NETWORK_FAMILY, inputPath: string
  * @param network
  * @param subgraphManifest
  */
-function graphManifestToSubqlManifest(
-  network: NETWORK_FAMILY,
-  subgraphManifest: SubgraphProject
-): CommonSubqueryProject {
+function graphManifestToSubqlManifest(chainInfo: ChainInfo, subgraphManifest: SubgraphProject): CommonSubqueryProject {
   return {
-    network: {chainId: subgraphManifest.dataSources[0].network, endpoint: ''},
-    name: subgraphManifest.description,
+    network: {chainId: chainInfo.chainId, endpoint: ''},
+    name: subgraphManifest.name,
     specVersion: '1.0.0',
+    runner: {
+      node: {version: '^', name: findRunnerByNetworkFamily(chainInfo.networkFamily)},
+      query: {version: '^', name: '@subql/query'},
+    },
     version: subgraphManifest.specVersion,
-    dataSources: subgraphDsToSubqlDs(network, subgraphManifest.dataSources),
+    dataSources: subgraphDsToSubqlDs(chainInfo.networkFamily, subgraphManifest.dataSources),
     description: subgraphManifest.description,
     schema: subgraphManifest.schema,
+    repository: subgraphManifest.repository,
     templates: subgraphManifest.templates
-      ? subgraphTemplateToSubqlTemplate(network, subgraphManifest.templates)
+      ? subgraphTemplateToSubqlTemplate(chainInfo.networkFamily, subgraphManifest.templates)
       : undefined,
   };
 }
@@ -80,38 +103,19 @@ export function subgraphTemplateToSubqlTemplate(
   subgraphTemplates: SubgraphTemplate[]
 ): TemplateKind[] {
   const convertFunction = networkTemplateConverters[network as NETWORK_FAMILY];
+  if (!convertFunction) {
+    throw new Error(`When migration, unable find convert template method for ${network} `);
+  }
   return subgraphTemplates.map((t) => convertFunction(t));
 }
 
-export function subgraphDsToSubqlDs(network: NETWORK_FAMILY, subgraphDs: SubgraphDataSource[]): DatasourceKind[] {
+export function subgraphDsToSubqlDs(
+  network: NETWORK_FAMILY,
+  subgraphDs: SubgraphDataSource[]
+): MigrateDatasourceKind[] {
   const convertFunction = networkDsConverters[network];
+  if (!convertFunction) {
+    throw new Error(`When migration, unable find convert dataSource method for ${network} `);
+  }
   return subgraphDs.map((ds) => convertFunction(ds));
 }
-
-// TODO, currently use DatasourceKind, which migrate network supported,should be a new type include all network dataSources
-type DsConvertFunction = (ds: SubgraphDataSource) => DatasourceKind;
-type TemplateConvertFunction = (ds: SubgraphTemplate) => TemplateKind;
-
-// @ts-ignore
-export const networkDsConverters: Record<NETWORK_FAMILY, DsConvertFunction> = {
-  [NETWORK_FAMILY.ethereum]: convertEthereumDs,
-  // [NETWORK_FAMILY.substrate]: convertSubstrateDs,
-  // [NETWORK_FAMILY.cosmos]: convertCosmosDs,
-  // [NETWORK_FAMILY.algorand]: convertAlgorandDs,
-  // [NETWORK_FAMILY.flare]: convertFlareDs,
-  // [NETWORK_FAMILY.near]: convertNearDs,
-  // [NETWORK_FAMILY.stellar]: convertStellarDs,
-  // [NETWORK_FAMILY.concordium]: convertConcordiumDs,
-};
-
-// @ts-ignore
-const networkTemplateConverters: Record<NETWORK_FAMILY, TemplateConvertFunction> = {
-  [NETWORK_FAMILY.ethereum]: convertEthereumTemplate,
-  // [NETWORK_FAMILY.substrate]: convertSubstrateDs,
-  // [NETWORK_FAMILY.cosmos]: convertCosmosDs,
-  // [NETWORK_FAMILY.algorand]: convertAlgorandDs,
-  // [NETWORK_FAMILY.flare]: convertFlareDs,
-  // [NETWORK_FAMILY.near]: convertNearDs,
-  // [NETWORK_FAMILY.stellar]: convertStellarDs,
-  // [NETWORK_FAMILY.concordium]: convertConcordiumDs,
-};