+++ packages/cli/src/utils/networkFamily.ts
@@ -1,7 +1,7 @@
 // Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import {NETWORK_FAMILY} from '@subql/common';
+import {NETWORK_FAMILY, runnerMapping} from '@subql/common';
 
 const lowerCaseFirst = (str: string) => str.charAt(0).toLowerCase() + str.slice(1);
 
@@ -15,4 +15,13 @@ export function getNetworkFamily(network: string): NETWORK_FAMILY {
     throw new Error(`Network not found or unsupported network ${network}`);
   }
   return NETWORK_FAMILY[lowerCaseFirst(network) as unknown as keyof typeof NETWORK_FAMILY];
+}
+
+export function findRunnerByNetworkFamily(networkFamily: NETWORK_FAMILY): string | undefined {
+  for (const [key, value] of Object.entries(runnerMapping)) {
+    if (value === networkFamily) {
+      return key;
+    }
+  }
+  return undefined;
 }